about to test degrade_votes method

+++ db/migrate/20140703145711_create_votes.rb
@@ -1,7 +1,8 @@
 class CreateVotes < ActiveRecord::Migration
   def change
       create_table :votes do |t|
-        t.integer :number
+        t.float :number
+        
         t.belongs_to :secret
     end
   end
